Separate each page in their own folder and components
