+++ src/views/MainPage/index.js
@@ -1,0 +1,2 @@
+import MainPage from "./MainPage";
+export default MainPage;
