+++ src/views/PostsPage/FilterBar.js
