+++ src/views/PostsPage/index.js
@@ -1,0 +1,2 @@
+import PostsPage from "./PostsPage";
+export default PostsPage;
