+++ src/views/SinglePostPage/SinglePostPage.js
