+++ src/views/SinglePostPage/index.js
@@ -1,0 +1,2 @@
+import SinglePostPage from "./SinglePostPage";
+export default SinglePostPage;
